UserManager/requirements.txt
Flask
grpcio
grpcio-tools
mysql-connector-python
python-dotenv
bcrypt